+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to `laravel-filament-flexible-content-blocks` will be documented in this file.
 
+## v2.7.1 - 2026-01-22
+
+- Make hero video url field in hero image section optional with create flag
+
+**Full Changelog**: https://github.com/statikbe/laravel-filament-flexible-content-blocks/compare/v2.7.0...v2.7.1
+
+## v2.7.0 - 2026-01-19
+
+- make hero changes more flexible, see FlexibleSelectField
+- video support in hero
+- add tests
+- fix phpstan errors
+
 ## v2.6.4 - 2025-10-21
 
 - Improved config return type for getCallToActionModels